organisation code

+++ src/Soldat.java
@@ -7,7 +7,7 @@ public class Soldat extends Unite{
 	int prioriteDEF = 1;
 	int mvtParTour = 2;
 	
-	public Soldat(int nombreUnites, int idJoueur, Double x, Double y) {
+	public Soldat(int nombreUnites, int idJoueur, int x, int y) {
 		super(nombreUnites, idJoueur);
 	}
